+++ firmware/quadruna/VC/CMakeLists.txt
@@ -32,7 +32,7 @@ list(APPEND IO_SRCS
         "${SHARED_IO_INCLUDE_DIR}/io_led.c"
         "${SHARED_IO_INCLUDE_DIR}/io_time.c"
         "${SHARED_IO_INCLUDE_DIR}/io_taskMonitor.c"
-        "${SHARED_IO_INCLUDE_DIR}/io_shared_imu.c"
+        "${SHARED_IO_INCLUDE_DIR}/io_imu.c"
 
 )
 set(IO_INCLUDE_DIRS "${CMAKE_CURRENT_SOURCE_DIR}/src/io" "${SHARED_IO_INCLUDE_DIR}" "${SHARED_IO_INCLUDE_QUADRUNA_DIR}")
