+++ firmware/quadruna/VC/src/io/io_imu_config.c
@@ -1,0 +1,3 @@
+#include "io_imu_config.h"
+
+const I2cDevice *imu_i2c_handle = &imu_i2c;
