+++ firmware/quadruna/VC/src/io/io_imu_config.h
@@ -1,0 +1,5 @@
+#include "hw_i2cs.h"
+#include "hw_i2c.h"
+#include "io_imu.h"
+
+extern const I2cDevice *imu_i2c_handle;
